pyproject.toml
[project]
name = "axiondeep-research"
version = "0.1.0"
description = "Axion Deep Labs — Experimental Research Programs"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",

    # Topology (EXP-01)
    "ripser>=0.6.4",
    "persim>=0.3.1",
    "scikit-tda>=1.1.0",

    # Tracking & Visualization
    "wandb>=0.16.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",

    # Config
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
